Implement route matching, allow custom 404 func
Before this commit, regex matching was not working properly. Basically,
it would match with the first fit and not the best fit. So, if the first
handler you add handles "/", all routes were resolving to that route.

Now we check all routes to see if they match. If no routes match, we
return `None` as before. If a route matches, we get the length of
the match and if it is longer than the current longest match, we store
its length and index.

There are still a number of corner cases that are not covered, but this
is functional for now.